Update bootstrap_count_matrix so that it can take non-integer matrices
Note that if the input X is an integer matrix, this modification will not change anything. Previously, when the input X is non-integer, the code will throw an error because x/N is not guaranteed to sum to 1. Now, even when the input X is non-integer, the code will not throw an error. But keep in mind that the output will always be integers.