Improve stratify_samples()
We added a silhouette score threshold after running OptimalK. The clustering (when k > 1) will be accepted if there is at least one cluster with silhouette score greater than the threshold (default 0.9), and rejected if there is not. This helps distinguish distinct clusters such as MMRD vs. MMRP from not-so-distinct clusters such as APOBEC vs. non-APOBEC.