Add final NNLS at the end of NMF and mvNMF
The motivation is that, with this final NNLS, the lambda tilde selection for mvNMF seems to be working better (see 20210715_figures.0_exploratory.key). We therefore add this final NNLS at the end of both NMF and mvNMF. We also normalize W along with that.

Note that the raw output from the NMF/mvNMF algorithm is now saved as self._W and self._H