Update sil_score definition for corner cases
There are six cases in total.
1. n_replicates = 1, n_replicates_after_filtering = 1, n_components = 1: In this case we want sil_score = 1, i.e., we don't want to filter out this solution based on stability.
2. n_replicates > 1, n_replicates_after_filtering = 1, n_components = 1: In this case we still set sil_score = 1. We won't filter out this solution based on sil_score. But we'll look at the ratio n_replicates_after_filtering/n_replicates in _select_n_components() to do appropriate filtering.
3. n_replicates > 1, n_replicates_after_filtering > 1, n_components = 1: Similar to 2.
4. n_replicates = 1, n_replicates_after_filtering = 1, n_components > 1: This means we are not running any replicates. So we won't do any filtering on stability. So we set sil_score to be 1.
5. n_replicates > 1, n_replicates_after_filtering = 1, n_components > 1: Similar to 2, we'll set sil_score to be 1.
6. n_replicates > 1, n_replicates_after_filtering > 1, n_components > 1: This is not a corner case. Sil_score will be calculated.

In summary:
1. When n_components = 1, we always set sil_score to be 1. No filtering will be done based on sil_score. Potential filtering will be done based on the ratio n_replicates_after_filtering/n_replicates.
2. When n_components > 1, and n_replicates_after_filtering = 1, we set sil_score to be 1. No filtering will be done based on sil_score. Potential filtering will be done based on the ratio n_replicates_after_filtering/n_replicates.
3. When n_components > 1 and n_replicates_after_filtering > 1, sil_score can be calculated, i.e., is well-defined.

Now, the only remaining cases when sil_score is not well-defined is when n_components > 1 and n_replicates_after_filtering > 1, and when hierarchical clustering is used. In this case, there could be clusters with only a single signature, in which case the sil_score will be set to be 0 by sklearn. That makes sense since if the cluster contains only a single signature, that means it is not a stable signature.

Of course, this assumes that n_replicates_after_filtering is sufficiently large, say, not 2. We'll only use sufficiently large n_replicates in our runs. So it should be fine. When n_replicates_after_filtering is too small, it will be captured by the ratio n_replicates/n_replicates_after_filtering, and when n_replicates_after_filtering is large enough, then sil_score will capture unstable solutions.

We'll mostly be using clustering_by_matching anyway. So we should be fine.