Change the default value of n_replicates_filter_ratio_thresh to 0.2.
We don't want the stability definition to depend too much on n_replicates_after_filtering/n_replicates.

Before we were using 0.5. That makes sense, since 'stable' perhaps means that at least the majority (i.e., > 50%) of the solutions are equivalently good.

But I observed that mvNMF solutions could be quite unstable, due to the fact that we only tune lambda_tilde for one of the mvNMF runs. That results in relatively lower ratio of n_replicates_after_filtering/n_replicates. So I'm lowering the threshold here.