Add an option to add noise to the samplewise errors in wrappedMVNMF
The motivation is that, in low-dimensional toy examples, when the reconstruction can be precise, the samplewise errors are very small (e.g., in the order of 1e-7), resulting in very small differences between different lambda_tilde values (e.g., in the order of 1e-14). However, this difference tends to be systematic, i.e., always shifted towards positive or negative values, resulting in significant p-values in rank sum test or differential tail test. These small difference should not be considered significant, but rather numerical errors.

We thus allow some noise to the samplewise errors in these cases.