Update DenovoSig to allow more control on _select_n_components explicitly.
Note that the repeated run of _select_n_components with different methods does not make sense anymore, since the method is now controlled by an explicit parameter. We'll need to separate out select_n_components as a separate function in DenovoSig, in which case we can easily run it multiple times, without the need to rerun NMF or mvNMF. We'll remove this redundancy when we do that.