Update OptimalK._cluster_statistic() such that it does not throw an error when k is the same as n_samples
Silhouette score can only be calculated when number of clusters is from 2 to n_samples - 1. We output nan for both 1 and n_samples.

Similar error catching may also need to be done in denovo._gather_results(). But let's not worry about it for now since there it won't really happen when number of clusters is 1 or n_samples, the former of which is taken care of separately, and the latter of which usually won't happen because n_replicates = 1 is taken care of separately.