Update OptimalK such that when no candidates are available, the greatest k is selected.
When candidates are empty based on gap statistic or log gap statistic, we select the greatest k. This can happen when max_k is equal to the number of samples. So basically now we can allow the optimal k to be equal to the number of samples. Note that when n_samples is indeed the best solution, only gap statistic works well. Log gap statistic does not. And sil score is not defined when k = 1 or n_samples. So sil score does not allow k = 1 or n_samples.