Add n_support output in _gather_results()
For hierarchical clustering, n_support is the number of signatures within the cluster corresponding to that signature. This number gives a sense of how many times this signature is discovered in the replicates. For example, if out of 20 replicates, only 2 of them have this signature, that may indicate that this signature is less likely to be real. Note that in this case the sil score could still be high.

Ideally this information should also be used when we define stable solutions. But right now we do not use it.