Critical update of match_signature_to_catalog
Previously, when doublets are considered, singlets are ignored, and when triplets are considered, doublets and singlets are ignored. This is undesirable when min_contribution > 0. Because when min_contribution > 0, it can happen that the best solution is a doublet instead of a triplet, i.e., only when min_contribution = 0, it is guaranteed that there exists a triplet that is better than all doublets.

So, we now modified the code such that when triplets are considered, doublets and singlets are still included in the comparison.

This update fixes undesirable results when min_contribution > 0.