Modify match in terms of connected_sigs
Now matching will always be performed with SparseNNLS with indices_associated_sigs set to None.

Then in the final W_s, we check what signatures are there, and add missing connected signatures if there are.

This is a better approach, because if we simply use indices_associated_sigs in SparseNNLS, then after _get_W_s() and _clear_W_s(), we may still lose connected signatures.