Add some error catching in validate and validate_grid
When mvNMF is used for de novo discovery, during validation, we fix the lambda_tilde as selected during mvNMF.

However, if mvnmf_hyperparameter_method is originally set to 'all', it becomes a bit more complicated. Current implementation does not take care of that gracefully. So we'll throw an error if mvnmf_hyperparameter_method is originally set to 'all' and when validation is performed.

This won't have a huge effect, as normally we don't set mvnmf_hyperparameter_method to 'all'.